+++ frontend/src/pages/dreams/mine/MyDreamPage.tsx
@@ -1,19 +1,19 @@
 import { fetchMyDreams } from "@/api/dreams/mine";
 import { LoadingContext } from "@/contexts/LoadingContext";
-import { Dream } from "@/types/dream";
+import { Dream, MyDreamSortKey } from "@/types/dream";
 import { useContext, useEffect, useState } from "react";
 import { useNavigate } from "react-router";
 import Header from "../components/Header";
 import SakuraScatterEffect from "../components/SakuraScatterEffect";
 import MyDreamCards from "./MyDreamCards";
-import MyDreamInput from "./MyDreamInput";
-
-export type MyDreamSortKey = "updated_at" | "likes";
+import MyDreamInput from "./form/MyDreamForm";
 
 const MyDreamPage = () => {
   const [myDreams, setMyDreams] = useState<Dream[]>([]);
-  const [sortKey, setSortKey] = useState<MyDreamSortKey>("updated_at");
+  // prettier-ignore
+  const [myDreamSortKey, setMyDreamSortKey] = useState<MyDreamSortKey>("updated_at");
   const { setIsLoading } = useContext(LoadingContext);
+
   const navigate = useNavigate();
 
   useEffect(() => {
@@ -26,7 +26,7 @@ const MyDreamPage = () => {
     const loadMyDreams = async () => {
       try {
         setIsLoading(true);
-        const dreams: Dream[] = await fetchMyDreams(sortKey);
+        const dreams: Dream[] = await fetchMyDreams(myDreamSortKey);
         setMyDreams(dreams);
       } catch (e) {
         console.error(e);
@@ -36,33 +36,18 @@ const MyDreamPage = () => {
     };
 
     loadMyDreams();
-  }, [sortKey]);
+  }, [myDreamSortKey]);
 
   return (
-    <div>
+    <>
       <Header />
       <SakuraScatterEffect />
-
-      <div className="p-4">
-        <MyDreamInput setMyDreams={setMyDreams} />
-
-        <div className="mt-3 ml-5 flex items-center gap-2">
-          <span className="text-gray-600">並び替え:</span>
-          <select
-            value={sortKey}
-            onChange={(event) => {
-              setSortKey(event.target.value as MyDreamSortKey);
-            }}
-            className="border rounded px-3 py-1 bg-white text-gray-800 shadow-sm"
-          >
-            <option value="updated_at">更新日順</option>
-            <option value="likes">いいね順</option>
-          </select>
-        </div>
-      </div>
-
-      <MyDreamCards myDreams={myDreams} setMyDreams={setMyDreams} />
-    </div>
+      <MyDreamInput setMyDreams={setMyDreams} />
+      <MyDreamCards
+        myDreams={{ myDreams, setMyDreams }}
+        sortKey={{ myDreamSortKey, setMyDreamSortKey }}
+      />
+    </>
   );
 };
